Rewrite function for sampling transcriptome
Instead of two functions for sampling genome mode and sampling transcriptome mode we use only one function for sample reads from reference. We rewrite sampling function to adjust it for RNA data. For now we sample the whole transcript (to make it comparable to squigulator) and use read strand +. This solves the former bug with total_l which was incorrectly calculated for transcriptome mode. Also read_check is slightly modified to adjust for RNA data, where we do not check for sampled read_length vs real read_length since we sample the whole transcript.