Add first padding mask implemnt idea
Added a first draft for a padding mask for the decoder part. This will set the padded signal points to -inf in the decoder self attention output. However this has negative impact on the performance since generating the mask is ineffective. Since it has no advantages in terms of runtime or memory consumption, this will maybe get discarded